dir int -->float
Lines 40 & 58

+++ include/linedetect.hpp
@@ -37,7 +37,7 @@ class LineDetect {
  public:
     cv::Mat img;  /// Input image in opencv matrix format
     cv::Mat img_filt;  /// Filtered image in opencv matrix format
-    int dir;  /// Direction message to be published
+    double dir;  /// Direction message to be published
 /**
 *@brief Callback used to subscribe to the image topic from the Turtlebot and convert to opencv image format
 *@param msg is the image message for ROS
@@ -55,7 +55,7 @@ class LineDetect {
 *@param input is the Filtered input image in opencv matrix format
 *@return int direction which returns the direction the turtlebot should head in
 */
-    int colorthresh(cv::Mat input);
+    double colorthresh(cv::Mat input);
 
  private:
     cv::Scalar LowerYellow;
